refactor: 응답 내부에만 쓰이는 DTO 내부 클래스로 이동
