--- src/main/java/kitchenpos/dto/response/MenuProductResponse.java
@@ -1,34 +1,0 @@
-package kitchenpos.dto.response;
-
-import java.util.List;
-import java.util.stream.Collectors;
-import kitchenpos.domain.MenuProduct;
-
-public class MenuProductResponse {
-
-    private final long productId;
-    private final long quantity;
-
-    private MenuProductResponse(final long productId, final long quantity) {
-        this.productId = productId;
-        this.quantity = quantity;
-    }
-
-    public static MenuProductResponse from(final MenuProduct menuProduct) {
-        return new MenuProductResponse(menuProduct.getProduct().getId(), menuProduct.getQuantity());
-    }
-
-    public static List<MenuProductResponse> of(final List<MenuProduct> menuProducts) {
-        return menuProducts.stream()
-                .map(MenuProductResponse::from)
-                .collect(Collectors.toList());
-    }
-
-    public long getProductId() {
-        return productId;
-    }
-
-    public long getQuantity() {
-        return quantity;
-    }
-}
