+++ src/main/java/kitchenpos/dto/response/MenuResponse.java
@@ -4,6 +4,7 @@ import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.domain.Menu;
+import kitchenpos.domain.MenuProduct;
 
 public class MenuResponse {
 
@@ -59,5 +60,34 @@ public class MenuResponse {
 
     public List<MenuProductResponse> getMenuProducts() {
         return menuProducts;
+    }
+
+    public static class MenuProductResponse {
+
+        private final long productId;
+        private final long quantity;
+
+        private MenuProductResponse(final long productId, final long quantity) {
+            this.productId = productId;
+            this.quantity = quantity;
+        }
+
+        public static MenuProductResponse from(final MenuProduct menuProduct) {
+            return new MenuProductResponse(menuProduct.getProduct().getId(), menuProduct.getQuantity());
+        }
+
+        public static List<MenuProductResponse> of(final List<MenuProduct> menuProducts) {
+            return menuProducts.stream()
+                    .map(MenuProductResponse::from)
+                    .collect(Collectors.toList());
+        }
+
+        public long getProductId() {
+            return productId;
+        }
+
+        public long getQuantity() {
+            return quantity;
+        }
     }
 }
